+++ opm/simulators/flow/FlowGenericProblem.hpp
@@ -52,13 +52,6 @@ class EclipseState;
 class Schedule;
 template<typename Grid, typename GridView> class LookUpData;
 
-int eclPositionalParameter(std::function<void(const std::string&,
-                                              const std::string&)> addKey,
-                           std::set<std::string>& seenParams,
-                           std::string& errorMsg,
-                           const char** argv,
-                           int paramIdx);
-
 /*!
  * \ingroup BlackOilSimulator
  *
