+++ opm/simulators/flow/FlowGenericProblem_impl.hpp
@@ -39,44 +39,10 @@
 #include <fmt/format.h>
 #include <fmt/ranges.h>
 
-#include <algorithm>
 #include <iostream>
 #include <stdexcept>
 
 namespace Opm {
-
-int eclPositionalParameter(std::function<void(const std::string&, const std::string&)> addKey,
-                           std::set<std::string>& seenParams,
-                           std::string& errorMsg,
-                           const char** argv,
-                           int paramIdx)
-{
-    std::string param  = argv[paramIdx];
-    std::size_t i = param.find('=');
-    if (i != std::string::npos) {
-        std::string oldParamName = param.substr(0, i);
-        std::string oldParamValue = param.substr(i+1);
-        std::string newParamName = "--" + oldParamName;
-        std::replace(newParamName.begin(),
-                     newParamName.end(), '_' , '-');
-        errorMsg =
-          "The old syntax to specify parameters on the command line is no longer supported: "
-          "Try replacing '" + oldParamName + "=" + oldParamValue + "' with "+
-          "'" + newParamName + "=" + oldParamValue + "'!";
-        return 0;
-    }
-
-    if (seenParams.count("EclDeckFileName") > 0) {
-        errorMsg =
-            "Parameter 'EclDeckFileName' specified multiple times"
-            " as a command line parameter";
-        return 0;
-    }
-
-    addKey("EclDeckFileName", argv[paramIdx]);
-    seenParams.insert("EclDeckFileName");
-    return 1;
-}
 
 template<class GridView, class FluidSystem>
 FlowGenericProblem<GridView,FluidSystem>::
